Update ex7_22.h
While the original (using 'struct') was technically correct, I believe using 'class' makes more sense here.  Here's what the book states: 
"As a matter of programming style, when we define a class intending for all of its
members to be public, we use struct. If we intend to have private members,
then we use class."